+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 - The help message on invalid command line arguments now shows an environment-specific path to the executable by using the implicit zeroth path argument.
+- Built-in functions no longer each have a custom declaration.
 
 ### Fixed
 - The error line is always `1`.